+++ NEWS.md
@@ -11,3 +11,4 @@
 
 - added NEWS.md creation with newsmd
 - added reference to a pre-print paper in bioRxiv 
+- removed changes to the global options of the user